requirements.txt
discord.py==2.4
openai==1.54.3
peewee==3.17.7
pydantic==2.9.2
pydantic-extra-types==2.10.0
python-dotenv==1.0.1
table2ascii==1.1.3
toml==0.10.2
tzdata==2024.2
